Moved ItemType/WidgetType to separate file ItemType.swift Taking advantage of enum with raw value of String to be Decodable while graciously handling of unknown enum types
